Move query reporting/deltas into notifications
